fish/requirements.txt
django==2.1.15
djangorestframework==3.11.0
PyMySQL==0.9.3
Pillow==7.1.2